pyproject.toml
[project]
name = "ai-materials"
version = "0.1.0"
description = "Repository for LLM experimentation, optimization, and evaluation with Jupyter notebooks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.25.0",
    "dspy-ai>=2.4.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "datasets>=2.14.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "python-dotenv>=1.2.1",
    "langfuse>=3.8.1",
    "dspy>=3.0.3",
    "openinference-instrumentation-dspy>=0.1.30",
]
